+++ CHANGELOG.md
@@ -59,6 +59,7 @@
   - allow both flow-matching and discrete samplers for sdxl models  
   - cleanup command line parameters  
   - add `--new` command line flag to enable testing of new packages without breaking existing installs  
+  - downgrade rocm to `torch==2.7.1`  
 - **OpenVINO**
   - update to `openvino==2025.3.0`  
   - add deprecation warning for Python 3.9